requirements.txt
openai>=1.3.0
python-dotenv
furhat-realtime-api
hume
websockets
